+++ Classes/CeImport/DataCollection/CeApiDataCollection.php
@@ -5,10 +5,13 @@ declare(strict_types=1);
 namespace BrainAppeal\CampusEventsConnector\CeImport\DataCollection;
 
 use BrainAppeal\CampusEventsConnector\CeImport\DataTransformer\DefaultDataTransformer;
+use BrainAppeal\CampusEventsConnector\CeImport\DataTransformer\EventDataTransformer;
+use BrainAppeal\CampusEventsConnector\Import\DataCollection\AbstractDataCollection;
 use BrainAppeal\CampusEventsConnector\Import\Exception\ApiLimitReachedException;
 use BrainAppeal\CampusEventsConnector\Import\Exception\ApiRecordNotFoundException;
-use BrainAppeal\CampusEventsConnector\Import\DataCollection\AbstractDataCollection;
+use BrainAppeal\CampusEventsConnector\Import\Exception\StopImportException;
 use BrainAppeal\CampusEventsConnector\Import\Model\ImportRecordModel;
+use BrainAppeal\CampusEventsConnector\Utility\TCAUtility;
 
 class CeApiDataCollection extends AbstractDataCollection
 {
@@ -16,6 +19,8 @@ class CeApiDataCollection extends AbstractDataCollection
      * If set to false, no further API calls will be made during this run, but the remaining records will be processed
      */
     protected bool $apiAccessEnabled = true;
+
+    protected bool $delayLoadingDetails = false;
     private CeApiConnector $apiConnector;
 
     /**
@@ -89,23 +94,95 @@ class CeApiDataCollection extends AbstractDataCollection
     protected function startProcessingForNewImport(): void
     {
         gc_enable();
+        $this->delayLoadingDetails = !$this->getImportOptions()->isForceUpdate();
+        $eventSourceIdList = $this->fetchAndAddEventItems();
+        /** @var ImportRecordModel[] $rows */
         $rows = [];
-        $apiConnector = $this->apiConnector;
-        foreach ($this->dataTransformerFactory->getAll() as $dataTransformer) {
-            /** @var DefaultDataTransformer $dataTransformer */
-            $dataFullyLoaded = $dataTransformer->isApiListItemContainsAllData();
-            foreach ($this->languageMap as $languageCode => $languageUid) {
-                $listItems = $apiConnector->fetchItemListForType($dataTransformer, $languageCode);
-                foreach ($listItems as $record) {
-                    $record['_language_id'] = $languageUid;
-                    $record['_language_code'] = $languageCode;
-                    $this->addImportRecordModel($record, $dataTransformer, $dataFullyLoaded, $rows);
+        if ($this->getImportOptions()->isForceUpdate()) {
+            foreach ($this->dataTransformerFactory->getAll() as $dataTransformer) {
+                if ($dataTransformer->getTable() === TCAUtility::TABLE_EVENTS) {
+                    continue;
                 }
+                /** @var DefaultDataTransformer $dataTransformer */
+                $dataTransformer->setValidEventSourceIds($eventSourceIdList);
+                $this->fetchAndAddRowsForDataTransformer($dataTransformer, $rows);
             }
+            gc_collect_cycles();
+            if (!empty($rows)) {
+                $this->importRecordWriter->addRows($rows);
+            }
+            return;
+        }
+        $importId = $this->getImportOptions()->getImportId();
+        $incompleteRows = $this->importRecordWriter->findIncompleteImportRows($importId, TCAUtility::TABLE_EVENTS);
+        if (empty($incompleteRows)) {
+            throw new StopImportException(sprintf('All event rows are marked as skipped for import id %d. Stopping import process.', $importId));
+        }
+        foreach ($this->dataTransformerFactory->getAll() as $dataTransformer) {
+            if ($dataTransformer->getTable() === TCAUtility::TABLE_EVENTS) {
+                continue;
+            }
+            /** @var DefaultDataTransformer $dataTransformer */
+            $dataTransformer->setValidEventSourceIds($eventSourceIdList);
+            $this->fetchAndAddRowsForDataTransformer($dataTransformer, $rows);
         }
         gc_collect_cycles();
         if (!empty($rows)) {
             $this->importRecordWriter->addRows($rows);
+        }
+        // Only load the record details for rows that are not skipped
+        $incompleteRows = $this->importRecordWriter->findIncompleteImportRows($importId);
+        if (!empty($incompleteRows)) {
+            $dummyRows = [];
+            foreach ($incompleteRows as $model) {
+                $this->fetchRemainingDataForModel($model, $dummyRows);
+            }
+            $this->importRecordWriter->update($incompleteRows);
+        }
+    }
+
+    /**
+     * Fetches and processes event items by utilizing a data transformer and appends them to the import record writer.
+     *
+     * This method retrieves event-related rows, transforms them using a specific data transformer, and prepares
+     * the source record IDs of these events for further use.
+     *
+     * @return array List of source record UIDs associated with the processed event items.
+     */
+    protected function fetchAndAddEventItems(): array
+    {
+        /** @var ImportRecordModel[] $rows */
+        $rows = [];
+        // Save the event rows first. If no event was changed, we can stop the import process here.
+        $eventDataTransformer = $this->dataTransformerFactory->getDataTransformerByTable(TCAUtility::TABLE_EVENTS);
+        /** @var EventDataTransformer $eventDataTransformer */
+        $this->fetchAndAddRowsForDataTransformer($eventDataTransformer, $rows);
+        $eventSourceIdList = [];
+        foreach ($rows as $model) {
+            $eventSourceIdList[] = $model->getSourceRecordUid();
+        }
+        $this->importRecordWriter->addRows($rows);
+        return $eventSourceIdList;
+    }
+
+    /**
+     * Adds rows for the specified data transformer by processing API list items for each language
+     * in the language map. Updates the rows array with import record models derived from the API data.
+     *
+     * @param DefaultDataTransformer $dataTransformer The data transformer used to process and transform API records.
+     * @param array &$rows The array to be updated with import record models created from the API data.
+     * @return void
+     */
+    protected function fetchAndAddRowsForDataTransformer(DefaultDataTransformer $dataTransformer, array &$rows): void
+    {
+        $dataFullyLoaded = $dataTransformer->isApiListItemContainsAllData();
+        foreach ($this->languageMap as $languageCode => $languageUid) {
+            $listItems = $this->apiConnector->fetchItemListForType($dataTransformer, $languageCode);
+            foreach ($listItems as $record) {
+                $record['_language_id'] = $languageUid;
+                $record['_language_code'] = $languageCode;
+                $this->addImportRecordModel($record, $dataTransformer, $dataFullyLoaded, $rows);
+            }
         }
     }
 
@@ -119,6 +196,22 @@ class CeApiDataCollection extends AbstractDataCollection
      */
     protected function tryFetchRemainingDataForModel(ImportRecordModel $model, array &$newRows): bool
     {
+        if ($this->delayLoadingDetails) {
+            return false;
+        }
+        return $this->fetchRemainingDataForModel($model, $newRows);
+    }
+
+    /**
+     * Loads full API data for the given import record model if required and supported.
+     * Ensures detailed data is loaded for the model based on its source type and API constraints.
+     *
+     * @param ImportRecordModel $model The import record model for which detailed API data may be loaded.
+     * @param ImportRecordModel[] $rows An array passed by reference to collect new rows for processing.
+     * @return bool True, if the full API data was successfully loaded and updated in the model, false otherwise.
+     */
+    private function fetchRemainingDataForModel(ImportRecordModel $model, array &$rows): bool
+    {
         // Import detailed data until limit is reached
         if ($this->apiAccessEnabled && !$model->getDataFullyLoaded()) {
             $refDataTransformer = $this->dataTransformerFactory->getDataTransformerByTable($model->getTargetTable());
@@ -130,7 +223,7 @@ class CeApiDataCollection extends AbstractDataCollection
                     $fullRecordData = array_merge($record, $fullRecordData);
                     $refDataTransformer->updateImportRecord($model, $fullRecordData);
                     $model->setDataFullyLoaded(true);
-                    $this->postProcessAfterModelAdded($model, $newRows);
+                    $this->postProcessAfterModelAdded($model, $rows);
                     return true;
                 }
                 // The record was apparently deleted in the API data