+++ Classes/Import/Repository/AbstractImportRowRepository.php
@@ -6,6 +6,7 @@ namespace BrainAppeal\CampusEventsConnector\Import\Repository;
 
 use Doctrine\DBAL\ArrayParameterType;
 use BrainAppeal\CampusEventsConnector\Import\Model\ImportRecordModel;
+use Doctrine\DBAL\Exception;
 use TYPO3\CMS\Core\Database\Connection;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Database\Query\QueryBuilder;
@@ -35,6 +36,86 @@ abstract readonly class AbstractImportRowRepository
             $queryBuilder->set('import_data', json_encode(['finished' => 1]));
         }
         $queryBuilder->executeStatement();
+    }
+
+    /**
+     * Creates an ImportRecordModel instance from the given import row.
+     *
+     * @param array<string, mixed> $row Array representing a database row containing data required to create an ImportRecordModel.
+     * @return ImportRecordModel The populated ImportRecordModel instance created from the given row.
+     */
+    final protected function createImportRecordModelFromRow(array $row): ImportRecordModel
+    {
+        $typedIdentifier = !empty($row['source_record_uid']) ? (int)$row['source_record_uid'] : (string)$row['source_record_identifier'];
+        $model = new ImportRecordModel(
+            (string)$row['import_data'],
+            (string)$row['source_type'],
+            $typedIdentifier,
+            (string)$row['data_hash'],
+        );
+        $model->updateFromArray($row);
+        return $model;
+    }
+
+    /**
+     * Finds and retrieves rows from the import row table that match specific conditions
+     * related to the provided import entry.
+     *
+     * @param int $importId The import entry object used to filter the rows based on its unique identifier.
+     * @param string|null $targetTable Optional import table name to filter the rows.
+     * @return array<ImportRecordModel> An array of ImportRecordModel objects representing the matching rows
+     *               with the required conditions.
+     * @throws Exception
+     */
+    final public function findIncompleteImportRows(int $importId, ?string $targetTable = null): array
+    {
+        $queryBuilder = $this->createQueryBuilderForImportRowTable($importId);
+        $queryBuilder->andWhere(
+        // Only process records where the full data have been loaded
+            $queryBuilder->expr()->eq('data_fully_loaded', $queryBuilder->createNamedParameter(
+                0,
+                Connection::PARAM_INT
+            ))
+        );
+        if ($targetTable) {
+            $queryBuilder->andWhere(
+                $queryBuilder->expr()->eq(
+                    'source_type',
+                    $queryBuilder->createNamedParameter($targetTable)
+                )
+            );
+        }
+        $queryBuilder->orderBy('sys_language_uid', 'ASC');
+        $queryBuilder->addOrderBy('priority', 'DESC');
+        $queryBuilder->addOrderBy('source_record_uid', 'ASC');
+        return $this->getImportRowsWithQueryBuilder($queryBuilder);
+    }
+
+    /**
+     * Retrieves import rows using the given QueryBuilder instance.
+     *
+     * @param QueryBuilder $queryBuilder The query builder instance used to execute the query and fetch rows.
+     * @return array An array of import record models containing data from the executed query.
+     * @throws Exception
+     */
+    final protected function getImportRowsWithQueryBuilder(QueryBuilder $queryBuilder): array
+    {
+        $importResult = $queryBuilder->executeQuery();
+        $importRows = [];
+        $emptyRows = [];
+        while ($row = $importResult->fetchAssociative()) {
+            $model = $this->createImportRecordModelFromRow($row);
+            if (!$model->hasData()) {
+                $emptyRows[] = $model;
+            } else {
+                $importRows[] = $model;
+            }
+        }
+        $importResult->free();
+        if (!empty($emptyRows)) {
+            $this->markRowsAsFinished(null, $emptyRows);
+        }
+        return $importRows;
     }
 
     /**
@@ -88,7 +169,7 @@ abstract readonly class AbstractImportRowRepository
      * @return QueryBuilder An array of ImportRecordModel objects representing the matching rows
      *               with the required conditions.
      */
-    protected function createQueryBuilderForImportRowTable(int $importId): QueryBuilder
+    protected function createQueryBuilderForImportRowTable(int $importId, bool $includeSkipped = false): QueryBuilder
     {
         $table = self::TABLE_IMPORT_ROW;
         /** @var QueryBuilder $queryBuilder */
@@ -104,15 +185,19 @@ abstract readonly class AbstractImportRowRepository
                     $importId,
                     Connection::PARAM_INT
                 )),
-                $queryBuilder->expr()->eq('import_skipped', $queryBuilder->createNamedParameter(
-                    0,
-                    Connection::PARAM_INT
-                )),
                 $queryBuilder->expr()->eq('deleted', $queryBuilder->createNamedParameter(
                     0,
                     Connection::PARAM_INT
                 )),
             );
+        if (!$includeSkipped) {
+            $queryBuilder->andWhere(
+                $queryBuilder->expr()->eq('import_skipped', $queryBuilder->createNamedParameter(
+                    0,
+                    Connection::PARAM_INT
+                ))
+            );
+        }
         return $queryBuilder;
     }